Crate martock

martock is a game.

Initialization

In no particular order, the following must be initialized before the main loop. See the module docs for more information on each.

World

Arbiter

Reality

Sentience, which interact with the world and may implement Body, Committer, and/or Life

Main Loop

Give committers the state of the world and bodies and ask for commits to change the world state.

Pass all commits through the arbiter which decides whether the commit should be applied to the world state.

Apply reality to all the bodies (physics and other laws of nature).

Update all the mutable world observers (who implement the life trait), such as the camera, sound effects, etc.

Modules

arbiter
arbiter defends the world from invalid modifications.

block
block is the atomic unit of the world.

body
body is anything with a physical representation in the world subject to reality.

camera
Camera is responsible for keeping track of the current viewing window. It is not coupled to the player and may move independently.

commit
commit is the atomic unit of change to the world state.

committer
committer is anything that modifies the world.

interactive
interactive is anything influenced by user input.

life
life is anything that changes state in response to the world and bodies.

point
point manages absolute world locations.

reality
reality is laws and mechanics of nature.

render
render renders the state of the world to screen.

tool
tool is a medium by which to modify the world.

world
world manages the object world state.