Struct martock::world::World

pub struct World { /* fields omitted */ }

World manages changing and accessing the state of the world, which is a three dimensional array of blocks which can vary in state and behavior. World is explicitly not responsible for determining the content of the world by itself; World should not modify the state of the world except in fulfillment of a user's request.

From the user's perspective there is no concept of world boundaries; initialization is a per-block property and blocks are not allocated contiguously.

Methods

impl World

fn new() -> Self

new creates a new world.

fn block(&self, x: i64, y: u8) -> Option<Block>

Returns a block if it is initialized. If the block is Block::Void, it will return None.

fn set_block(&mut self, x: i64, y: u8, next_state: Block)

Sets a block state to the passed next_state. To deinitialize a Block, set the block state to Block::Void.

fn iter(&self, x: i64, y: u8, width: usize, height: usize) -> BlockIter